+++ ntex/src/ws/client.rs
@@ -457,6 +457,7 @@ impl WsConnection<Sealed> {
                 DispatchItem::Control(_) => Ok(None),
                 DispatchItem::Stop(Reason::KeepAliveTimeout) => Err(WsError::KeepAlive),
                 DispatchItem::Stop(Reason::ReadTimeout) => Err(WsError::ReadTimeout),
+                DispatchItem::Stop(Reason::WriteTimeout) => Err(WsError::WriteTimeout),
                 DispatchItem::Stop(Reason::Decoder(e) | Reason::Encoder(e)) => {
                     Err(WsError::Protocol(e))
                 }